+++ R/MODIS-package.R
@@ -14,7 +14,7 @@
 #' \cr
 #' \emph{Maintainer:} Florian Detsch \email{fdetsch@@web.de}
 #'
-#' @import bitops mapdata parallel ptw raster rgdal sp
+#' @import bitops mapdata parallel ptw raster rgdal sf sp
 #' @importFrom curl curl curl_download handle_setopt new_handle
 #' @importFrom devtools install_github
 #' @importFrom grDevices dev.new dev.off png
@@ -24,7 +24,6 @@
 #' @importFrom maptools map2SpatialPolygons map2SpatialLines checkPolygonsHoles
 #' @importFrom methods as new slot
 #' @importFrom rgeos gBuffer gIsValid
-#' @importFrom sf st_as_sf st_bbox
 #' @importFrom stats na.omit setNames smooth.spline
 #' @importFrom utils capture.output download.file installed.packages read.csv read.table vi write.csv write.table
 #' @rawNamespace if (.Platform$OS.type=="windows") importFrom(utils,shortPathName)